Time: 1 ms (86.67%), Space: 40.9 MB (68.12%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -1,19 +1,20 @@
 class Solution {
-    public int mySqrt(int x) {
-        int left = 1;
-        int right = x + 1;
+    public int mySqrt(long x) {
+        
+        long left = 1;
+        long right = x + 1;
 
         while(left < right) {
-            int mid = left + (right - left)/2;
+            long mid = left + (right-left)/2;
 
             if(mid*mid <= x) {
-                left = mid + 1;
+                left = mid+1;
             } else {
                 right = mid;
             }
-        }
+        } 
 
-        return left - 1;
+        return (int)left - 1;
         
     }
 }
